requirements.txt
pytest==9.0.2
pytest-order==1.3.0
pytest-html==3.2.0
playwright==1.55.0
pydantic==2.12.5
openpyxl
python-dotenv==1.2.1
pytest-xdist==3.8.0
allure-pytest==2.15.2
pytest-check
pytest-asyncio
